SiDECLARAv003_2_3_lts_instalacion/SiDECLARAv001_240221/proyecto/declaraciones/declaracion/admin.py
from .models import InfoPersonalFija
from .models.catalogos import CatEntesPublicos
from django.contrib.auth.models import User
from django.contrib.auth.admin import UserAdmin
from django.contrib import admin
from import_export import resources, widgets, fields
from import_export.widgets import ManyToManyWidget, ForeignKeyWidget
from import_export.admin import ImportExportModelAdmin,ExportMixin
from import_export.formats import base_formats


class UserResource(resources.ModelResource):
    class Meta:
        model = User
        fields = ('username','first_name', 'last_name', 'email','is_active','is_staff')

class UserAdmin(ImportExportModelAdmin, UserAdmin):
    """
    Class UserAdmin utliza librera para exportar e importar datos de un modelo a un excel
    """
    resource_class = UserResource
    pass

class FijaResourceAdmin(resources.ModelResource):
    class Meta:
        model = InfoPersonalFija

class adminFija(ImportExportModelAdmin):
    resource_class = FijaResourceAdmin
    list_display = ('nombre_completo','cat_puestos','usuario')
    exclude = ['owner','puesto', 'usuario', 'nombre_ente_publico', 
               'fecha_nacimiento','curp',
               'cat_entidades_federativas','cat_pais',
               'otro_ente', 'sended']
class CatEntesPublicosAdmin(resources.ModelResource):
    class Meta:
        model = CatEntesPublicos

class CatInstitucionResource(ImportExportModelAdmin):
    resource_class = CatEntesPublicosAdmin
    list_display = ('ente_publico','codigo')


admin.site.unregister(User)
admin.site.register(User, UserAdmin)
admin.site.register(InfoPersonalFija, adminFija)
admin.site.register(CatEntesPublicos,CatInstitucionResource)