SiDECLARAv003_2_3_lts_instalacion/SiDECLARAv001_240221/proyecto/declaraciones/requirements.txt
amqp==2.5.2
asgiref==3.2.3
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
backports.csv==1.0.7
beautifulsoup4==4.8.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.6.16
cffi==1.14.0
chardet==3.0.4
colorama==0.4.1
cryptography==2.8
defusedxml==0.6.0
diff-match-patch==20181111
distro==1.4.0
Django==3
django-anymail==7.0.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.0.1
django-easy-pdf==0.1.1
django-environ==0.4.5
django-excel==0.0.10
django-import-export==2.0
django-js-asset==1.2.2
django-mptt==0.10.0
django-multiselectfield==0.1.8
django-oauth-toolkit==1.2.0
django-redis-cache==2.1.0
django-smart-selects==1.5.6
djangorestframework==3.11.0
djangorestframework-simplejwt
entrypoints==0.3
et-xmlfile==1.0.1
flake8==3.7.9
html5lib==1.0.1
idna==2.8
importlib-metadata==0.19
isort==4.3.21
jdcal==1.4.1
kombu==4.6.6
lazy-object-proxy==1.4.1
lml==0.0.9
MarkupPy==1.14
mccabe==0.6.1
merge==1.0.0
more-itertools==7.2.0
mysql-connector==2.2.9
mysql-connector-python==8.0.17
mysqlclient==1.4.6
nltk==3.4.5
numpy==1.22.4
oauthlib==3.1.0
odfpy==1.4.0
openpyxl==3.0.10
packaging==19.1
pandas==1.3.5
pdfminer.six==20181108
pdfminer2==20151206
pdfminer3k==1.3.4
pdfrw==0.4
pdfx==1.3.0
Pillow==6.1.0
pluggy==0.12.0
ply==3.11
protobuf==3.11.2
psycopg2-binary==2.8.4
py==1.8.0
pycodestyle==2.5.0
pycparser==2.20
pycryptodome==3.8.2
pyexcel==0.5.15
pyexcel-io==0.5.20
pyexcel-webio==0.1.4
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.4.2
PyPDF2==1.26.0
PyPDF4==1.27.0
pytest==5.4.3
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
redis==3.3.11
rejson==0.5.4
reportlab==3.5.32
requests==2.22.0
simplekml==1.3.1
six==1.12.0
sortedcontainers==2.1.0
soupsieve==1.9.3
sqlparse==0.3.0
tablib==0.13.0
tabula==1.0.5
tabula-py==1.4.1
texttable==1.6.2
tika==1.19
tqdm==4.36.1
typed-ast==1.4.0
Unipath==1.1
urllib3==1.25.3
uWSGI==2.0.18
vine==1.3.0
virtualenv==16.7.2
virtualenvwrapper-win==1.2.5
wcwidth==0.1.7
webencodings==0.5.1
wrapt==1.11.2
xhtml2pdf==0.2.3
xlrd==1.2.0
xlwt==1.3.0
zipp==0.5.2
django-front==0.5.15
django-redis==5.2.0
django-colorfield==0.7.0
django-cors-middleware==1.5.0
WeasyPrint==52.5
